Test backend timestamp properties outside views
